+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## 0.2.0 - 2020-12-17
+
+
+### Fixed
+
+- Ability to see blueprint request button without permissions
+- Permission error viewing blueprint requests
+
 ## 0.2.0 - 2020-12-17
 
 ### Added